+++ src/YammerShell/YammerObjects/YammerFile.cs
@@ -4,10 +4,10 @@ namespace YammerShell.YammerObjects
 {
     public class YammerFile
     {
-        public int Id { get; set; }
-        public int NetworkId { get; set; }
-        public int? GroupId { get; set; }
-        public int? OwnerId { get; set; }
+        public long Id { get; set; }
+        public long NetworkId { get; set; }
+        public long? GroupId { get; set; }
+        public long? OwnerId { get; set; }
         public string Url { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
